Sentiment analysis model
Keras model with a tensorflow backend to analyze sentiment of IMDB
movie reviews.  Based on the usage of top 10000 words, a review is
classified as positive or negative.  Model outputs a probability
for sentiment.  A value less then 0.5 means the sentiment is negative
and a value >= 0.5 means the sentiment is postive.